chore: setup base for agent evals
