+++ package.json
@@ -58,6 +58,7 @@
     "@nestjs/schedule": "^5.0.1",
     "@nestjs/sequelize": "^11.0.0",
     "@slack/web-api": "^7.8.0",
+    "agentevals": "^0.0.5",
     "cache-manager": "4.1.0",
     "cache-manager-redis-store": "2.0.0",
     "lodash": "^4.17.21",
